fix: make repcli load runtime settings on demand
