+++ repcli.py
@@ -2,11 +2,14 @@ import glob
 import json
 import logging
 import os
+import sys
+import types
+from dataclasses import dataclass
 from datetime import date, datetime, timedelta
 from io import BytesIO
 from os.path import basename, getmtime, join
 from pathlib import Path
-from typing import Annotated
+from typing import TYPE_CHECKING, Annotated, Any
 
 import click
 import climage
@@ -18,12 +21,11 @@ from rich.logging import RichHandler
 from rich.table import Table
 from rich.theme import Theme
 
-from config import config
 from src.ai.utils import force_valid_json_string
-from src.persistence.replay_store import PlayerInfo, get_replay_store
-from src.playerinfo import save_player_info
-from src.replays.reader import ReplayReader
-from src.replays.types import Replay
+from src.runtime.settings import SettingsLoaderError, load_current_settings
+
+if TYPE_CHECKING:
+    from src.runtime.settings import Config
 
 custom_theme = Theme(
     {
@@ -38,8 +40,60 @@ log = logging.getLogger()
 log.setLevel(logging.WARNING)
 log.addHandler(RichHandler(show_time=False, rich_tracebacks=True))
 
-reader = ReplayReader()
-replay_store = get_replay_store()
+
+@dataclass
+class RepCliRuntime:
+    settings: "Config"
+    reader: Any
+    replay_store: Any
+    replay_model: type
+    player_info_model: type
+    save_player_info: Any
+
+
+def load_runtime_settings() -> "Config":
+    return load_current_settings()
+
+
+def _install_legacy_config(settings: "Config") -> None:
+    legacy_config = types.ModuleType("config")
+    legacy_config.config = settings
+    legacy_config.Config = type(settings)
+    legacy_config.SettingsLoaderError = SettingsLoaderError
+    legacy_config.load_current_settings = load_current_settings
+    sys.modules["config"] = legacy_config
+
+
+def _build_runtime() -> RepCliRuntime:
+    settings = load_runtime_settings()
+    _install_legacy_config(settings)
+
+    from src.persistence.database import get_database, set_database
+    from src.persistence.replay_store import PlayerInfo, get_replay_store
+    from src.playerinfo import save_player_info
+    from src.replays.reader import ReplayReader
+    from src.replays.types import Replay
+
+    database = get_database(settings)
+    set_database(database)
+    replay_store = get_replay_store(database)
+
+    return RepCliRuntime(
+        settings=settings,
+        reader=ReplayReader(),
+        replay_store=replay_store,
+        replay_model=Replay,
+        player_info_model=PlayerInfo,
+        save_player_info=save_player_info,
+    )
+
+
+def _get_runtime(ctx) -> RepCliRuntime:
+    runtime = ctx.obj.get("RUNTIME")
+    if runtime is None:
+        runtime = _build_runtime()
+        ctx.obj["RUNTIME"] = runtime
+    return runtime
 
 
 def dformat(x):
@@ -151,15 +205,19 @@ def cli(ctx, clean, debug, simulation, verbose):
 )
 def sync(ctx, from_: datetime, to_: datetime, from_most_recent: bool):
     """Sync replays and players from replay folder to MongoDB"""
+    runtime = _get_runtime(ctx)
+    settings = runtime.settings
 
     if from_most_recent:
         try:
-            most_recent = replay_store.get_most_recent_for_player(config.student.name)
+            most_recent = runtime.replay_store.get_most_recent_for_player(
+                settings.student.name
+            )
             sync_from_date = most_recent.unix_timestamp
         except ValueError:
             sync_from_date = from_.timestamp()
             console.print(
-                f":warning: No replays found in DB for {config.student.name}, falling back to --from"
+                f":warning: No replays found in DB for {settings.student.name}, falling back to --from"
             )
     else:
         sync_from_date = from_.timestamp()
@@ -167,7 +225,7 @@ def sync(ctx, from_: datetime, to_: datetime, from_most_recent: bool):
     sync_to_date = (to_ + timedelta(days=1)).timestamp()
 
     console.print(f"Syncing from {dformat(sync_from_date)} to {dformat(sync_to_date)}")
-    list_of_files = glob.glob(join(config.replay_folder, "*.SC2Replay"))
+    list_of_files = glob.glob(join(settings.replay_folder, "*.SC2Replay"))
     list_of_files = [
         f
         for f in list_of_files
@@ -177,39 +235,39 @@ def sync(ctx, from_: datetime, to_: datetime, from_most_recent: bool):
     list_of_files.sort(key=getmtime)
     console.print(f"Found {len(list_of_files)} potential replays to sync")
 
-    summary = _sync(list_of_files, ctx)
+    summary = _sync(list_of_files, ctx, runtime)
 
     console.print(summary.to_table())
 
 
-def _sync(list_of_files, ctx) -> SyncSummary:
+def _sync(list_of_files, ctx, runtime: RepCliRuntime) -> SyncSummary:
     summary = SyncSummary()
     summary.total_replays = len(list_of_files)
 
     for file_path in list_of_files:
-        replay_raw = reader.load_replay_raw(file_path)
-        if reader.apply_filters(replay_raw):
+        replay_raw = runtime.reader.load_replay_raw(file_path)
+        if runtime.reader.apply_filters(replay_raw):
             console.print(f"Adding {basename(file_path)}")
-            replay = reader.to_typed_replay(replay_raw)
+            replay = runtime.reader.to_typed_replay(replay_raw)
 
-            syncreplay(ctx, replay, summary)
-            syncplayer(ctx, replay, summary)
+            syncreplay(ctx, replay, summary, runtime)
+            syncplayer(ctx, replay, summary, runtime)
         else:
             console.print(f"Filtered {basename(file_path)}")
             summary.filtered_replays += 1
-            if reader.is_archon_mode(replay_raw):
+            if runtime.reader.is_archon_mode(replay_raw):
                 console.print(
                     f":couple: Archon mode is not supported {basename(file_path)}"
                 )
                 continue
-            if reader.is_instant_leave(replay_raw):
+            if runtime.reader.is_instant_leave(replay_raw):
                 summary.instant_leave_replays += 1
                 if ctx.obj["CLEAN"]:
                     summary.deleted_replays += 1
                     os.remove(file_path)
                     console.print(f":litter_in_bin_sign: Deleted {basename(file_path)}")
                     continue
-            if reader.has_afk_player(replay_raw):
+            if runtime.reader.has_afk_player(replay_raw):
                 summary.afk_replays += 1
                 if ctx.obj["CLEAN"]:
                     summary.deleted_replays += 1
@@ -224,12 +282,13 @@ def _sync(list_of_files, ctx) -> SyncSummary:
 def validate(logfile):
     "Validate all replays in the DB. Shows replays in DB which can't be read"
 
+    runtime = _get_runtime(click.get_current_context())
     summary = ValidationSummary()
 
-    for replay in replay_store.find_many_dict(Replay, raw_query={}):
+    for replay in runtime.replay_store.find_many_dict(runtime.replay_model, raw_query={}):
         console.print(f"Validating {basename(replay['filename'])}", end=" ")
         try:
-            rep = Replay(**replay)  # noqa: F841
+            rep = runtime.replay_model(**replay)  # noqa: F841
             console.print(":white_heavy_check_mark:")
             summary.valid_replays += 1
 
@@ -259,10 +318,13 @@ def query(ctx):
 @click.option("--query", "-q", type=str, prompt="MongoDB Query")
 def players(ctx, query):
     """Query the DB for players"""
+    runtime = _get_runtime(ctx)
     query = force_valid_json_string(query)
     query = json.loads(query)
 
-    players = replay_store.db.find_many(PlayerInfo, raw_query=query)
+    players = runtime.replay_store.db.find_many(
+        runtime.player_info_model, raw_query=query
+    )
 
     if not isinstance(players, list):
         console.print(f":x: Expected list of players, got {type(players).__name__}")
@@ -279,6 +341,7 @@ def players(ctx, query):
 @click.argument("replay", type=click.Path(exists=False), required=True, nargs=-1)
 def add(ctx, replay):
     """Add one or more replays to the DB"""
+    runtime = _get_runtime(ctx)
 
     list_of_files = []
 
@@ -295,7 +358,7 @@ def add(ctx, replay):
         console.print(f":x: No replays found for {replay}")
         return
 
-    summary = _sync(list_of_files, ctx)
+    summary = _sync(list_of_files, ctx, runtime)
 
     console.print(summary.to_table())
 
@@ -306,14 +369,15 @@ def add(ctx, replay):
 @click.option("--json", "-j", is_flag=True, default=False, help="Print in JSON")
 def echo(ctx, replay, json):
     """Echo pretty-printed parsed replay data from a .SC2Replay file"""
-    rep = reader.load_replay(replay)
+    runtime = _get_runtime(ctx)
+    rep = runtime.reader.load_replay(replay)
     if json:
         console.print_json(rep.default_projection_json())
     else:
         console.print(rep)
 
 
-def print_player_portrait(player: PlayerInfo):
+def print_player_portrait(player):
     portrait = player.portrait or player.portrait_constructed
     if portrait:
         img = Image.open(BytesIO(portrait)).resize((40, 40))
@@ -324,13 +388,13 @@ def print_player_portrait(player: PlayerInfo):
         print(output)
 
 
-def syncplayer(ctx, replay: Replay, summary: SyncSummary):
-    opponent = replay.get_opponent_of(config.student.name)
+def syncplayer(ctx, replay, summary: SyncSummary, runtime: RepCliRuntime):
+    opponent = replay.get_opponent_of(runtime.settings.student.name)
 
     if ctx.obj["SIMULATION"]:
         console.print(f"Simulation, would add {opponent.name} ({opponent.toon_handle})")
     else:
-        result, player_info = save_player_info(replay)
+        result, player_info = runtime.save_player_info(replay)
 
         if result.acknowledged:
             console.print(
@@ -351,11 +415,11 @@ def syncplayer(ctx, replay: Replay, summary: SyncSummary):
             )
 
 
-def syncreplay(ctx, replay: Replay, summary: SyncSummary):
+def syncreplay(ctx, replay, summary: SyncSummary, runtime: RepCliRuntime):
     if ctx.obj["SIMULATION"]:
         console.print(f"Simulation, would add {replay}")
     else:
-        result = replay_store.upsert(replay)
+        result = runtime.replay_store.upsert(replay)
         if result.acknowledged:
             console.print(f":white_heavy_check_mark: {replay} added to DB")
             summary.replays_added += 1
